requirements.txt
openai-agents>=0.0.9
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.6
gspread>=6.2.0
requests
jsonschema
streamlit>=1.24.0
google-cloud-aiplatform
google-generativeai
python-dotenv